feat(game): 实现游戏开始功能和地图数据优化
- 添加 ScheduledExecutorService 配置用于定时任务处理
- 将地图数据类型从 int[][] 改为 byte[][] 以优化内存使用
- 在 GameMatch 实体中添加 getMapData 方法用于获取字节字符串格式的地图数据
- 添加 START_GAME 消息类型用于游戏开始通知
- 在 GameMatchManager 中添加房间状态检查防止非等待状态房间启动游戏
- 实现 startGame 方法创建对战、推送开始游戏消息并设置房间状态
- 添加 LOADED 状态用于标识玩家已加载游戏资源
- 添加 StartNotice 消息格式用于传输房间ID、对战ID和地图数据

+++ src/main/java/com/zunf/tankbattletcpserver/config/MapConfig.java
@@ -35,8 +35,8 @@ public class MapConfig {
     private int spawnMargin = 2;
 
     // 地图元素编码
-    private int empty = 0;   // 空地
-    private int wall = 1;    // 不可破坏墙
-    private int brick = 2;   // 可破坏砖块
-    private int spawn = 3;   // 出生点
+    private byte empty = 0;   // 空地
+    private byte wall = 1;    // 不可破坏墙
+    private byte brick = 2;   // 可破坏砖块
+    private byte spawn = 3;   // 出生点
 }
